requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
easydict==1.9
filelock==3.7.0
fire==0.4.0
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
huggingface-hub==0.6.0
hydra-core==1.1.0
idna==3.3
imageio==2.19.1
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==2.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kornia==0.5.0
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pep8==1.7.1
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.8.0
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.3
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sahi==0.9.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.8.0
Shapely==1.8.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminaltables==3.1.10
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.2
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.1
typing-extensions==4.2.0
urllib3==1.26.9
webdataset==0.2.5
Werkzeug==2.1.2
wldhx.yadisk-direct==0.0.6
wrapt==1.14.1
yapf==0.32.0
yarl==1.7.2
zipp==3.8.0
